Update tests to properly(?) mock ServerService
Tests previously mocked ServerModel, which still required the Mongoose
module(s) to be loaded into the test as well - not isolated enough.
The tests now mock the entire ServerService instead, avoiding the issue.

The controller tests now test that the ServerController calls the
appropriate methods on the ServerService, though I'm fairly certain
there are better ways to test this as well.

Overall this is hopefully a good early example of separating out some
dependencies when testing, but with obvious drawbacks.

Clearly, separate tests should be written to test ServerService.